services/vision-agent/requirements.txt
flask==3.0.0
requests==2.31.0
psycopg2-binary==2.9.9
gunicorn==21.2.0
Pillow==10.2.0
base64==0.1.0
